chore: move renderWithAuth to test_hepers/index.tsx
