+++ site/src/test_helpers/index.tsx
@@ -2,9 +2,11 @@ import ThemeProvider from "@material-ui/styles/ThemeProvider"
 import { render as wrappedRender, RenderResult } from "@testing-library/react"
 import { createMemoryHistory } from "history"
 import React from "react"
-import { unstable_HistoryRouter as HistoryRouter } from "react-router-dom"
+import { MemoryRouter, Route, Routes, unstable_HistoryRouter as HistoryRouter } from "react-router-dom"
+import { RequireAuth } from "../components"
 import { dark } from "../theme"
 import { XServiceProvider } from "../xServices/StateContext"
+import { MockUser } from "./entities"
 
 export const history = createMemoryHistory()
 
@@ -20,6 +22,25 @@ export const WrapperComponent: React.FC = ({ children }) => {
 
 export const render = (component: React.ReactElement): RenderResult => {
   return wrappedRender(<WrapperComponent>{component}</WrapperComponent>)
+}
+
+type RenderWithAuthResult = RenderResult & { user: typeof MockUser }
+
+export function renderWithAuth(ui: JSX.Element, { route = "/" }: { route?: string } = {}): RenderWithAuthResult {
+  const renderResult = render(
+    <MemoryRouter initialEntries={[route]}>
+      <XServiceProvider>
+        <Routes>
+          <Route path={route} element={<RequireAuth>{ui}</RequireAuth>} />
+        </Routes>
+      </XServiceProvider>
+    </MemoryRouter>,
+  )
+
+  return {
+    user: MockUser,
+    ...renderResult,
+  }
 }
 
 export * from "./entities"
